requirements.txt
Pillow==9.0.1
numpy==1.21.2